requirements_video.txt
# Below for UMT and InternVideo2
transformers
datasets
scipy
sentencepiece
gdown
tqdm
ftfy
regex
git+https://github.com/openai/CLIP.git
open-clip-torch
opencv-python
opencv-python-headless
openai
einops
pandas
av
numpy
Pillow
timm
decord
huggingface-hub
wandb
matplotlib-inline
omegaconf
protobuf
pytz
requests
pyyaml
ftfy
easydict
regex
joblib
gitpython
einops
imageio
sentencepiece
# Below for InternVideo2
apex==0.9.10dev
deepspeed
fvcore
librosa
opencv_python
Pillow
psutil
PyYAML
soundfile
termcolor
wandb
packaging
ninja
peft
# Below are for LanguageBind
accelerate
evaluate
iopath
pytorchvideo @ git+https://github.com/linzhiqiu/pytorchvideo.git
tensorboardX
tokenizers
braceexpand
webdataset
urllib3